+++ packages/suite/src/constants/suite/urls.ts
@@ -45,7 +45,7 @@ export const WIKI_XPUB_URL =
     'https://wiki.trezor.io/User_manual:Displaying_account_public_key_(XPUB)';
 export const WIKI_BECH32_URL = 'https://wiki.trezor.io/Bech32';
 export const WIKI_P2SH_URL = 'https://wiki.trezor.io/P2SH';
-export const WIKI_P2PHK_URL = 'https://wiki.trezor.io/P2PHK_address';
+export const WIKI_P2PKH_URL = 'https://wiki.trezor.io/P2PKH';
 export const WIKI_ADVANCED_RECOVERY = 'https://wiki.trezor.io/User_manual:Advanced_recovery';
 export const WIKI_UDEV_RULES = 'https://wiki.trezor.io/Udev_rules';
 export const WIKI_TOR = 'https://wiki.trezor.io/Glossary:Tor';